#3: Use unsiged int instead of unsigned short for indices data type, because of the primitive restart value

+++ netcdf_to_gltf_converter/geometries.py
@@ -76,4 +76,4 @@ class TriangularMesh:
             np.ndarray: A two-dimensional numpy array with data type 'uint16'.
         """
         triangles = [triangle.as_list() for triangle in self.triangles]
-        return np.array(triangles, dtype="uint16")
+        return np.array(triangles, dtype="uint32")
